+++ cloud/objstore/objutil/chunk_reader.go
@@ -17,8 +17,8 @@ type ChunkReader struct {
 }
 
 // NewChunkReader creates a new chunk reader which will read chunks of the given size from the provided reader.
-func NewChunkReader(reader ioiface.ReadAtSeeker, chunkSize int64) ChunkReader {
-	return ChunkReader{size: chunkSize, reader: reader}
+func NewChunkReader(reader ioiface.ReadAtSeeker, size int64) ChunkReader {
+	return ChunkReader{size: size, reader: reader}
 }
 
 // ForEach breaks the 'reader' into chunks running the given function for each chunk created.
@@ -28,8 +28,17 @@ func (c ChunkReader) ForEach(fn func(chunk *io.SectionReader) error) error {
 		return fmt.Errorf("failed to determine length of reader: %w", err)
 	}
 
-	for s, e := int64(0), c.size-1; s < length; s, e = s+c.size, e+c.size {
-		if err := fn(io.NewSectionReader(c.reader, s, maths.Min(e+1, length)-s)); err != nil {
+	err = chunk(length, c.size, func(start, end int64) error {
+		return fn(io.NewSectionReader(c.reader, start, maths.Min(end+1, length+1)-start))
+	})
+
+	return err
+}
+
+// chunk runs the provided function creating 'size' chunks from zero to 'length'.
+func chunk(length, size int64, fn func(start, end int64) error) error {
+	for start, end := int64(0), size-1; start < length; start, end = start+size, end+size {
+		if err := fn(start, maths.Min(end, length-1)); err != nil {
 			return err
 		}
 	}